refactor: move utilites into utils/
