+++ lua/mcphub/utils/errors.lua
